requirements.txt
allure-behave         2.14.3
allure-python-commons 2.14.3
attrs                 25.3.0
behave                1.2.6
greenlet              3.2.3
parse                 1.20.2
parse_type            0.6.4
pip                   25.1.1
playwright            1.52.0
pluggy                1.6.0
pyee                  13.0.0
python-dotenv         1.1.0
setuptools            65.5.0
six                   1.17.0
typing_extensions     4.14.0


#./test_runner.sh all
#./test_runner.sh all
#./test_runner.sh 
